Remove examples api
