--- sites/svelte.dev/src/routes/+page.js
@@ -1,1 +1,0 @@
-export const prerender = true;
