+++ sites/svelte.dev/src/routes/_components/Demo.svelte
@@ -1,34 +1,11 @@
 <script>
+	import { getContext } from 'svelte';
 	import Example from './Example.svelte';
 	import Section from './Section.svelte';
 
-	const examples = [
-		{
-			id: 'hello-world',
-			title: 'Hello World',
-			description: 'Svelte components are built on top of HTML. Just add data.',
-		},
-		{
-			id: 'nested-components',
-			title: 'Scoped CSS',
-			description:
-				'CSS is component-scoped by default — no more style collisions or specificity wars. Or you can <a href="/blog/svelte-css-in-js">use your favourite CSS-in-JS library</a >.',
-		},
-		{
-			id: 'reactive-assignments',
-			title: 'Reactivity',
-			description:
-				'Trigger efficient, granular updates by assigning to local variables. The compiler does the rest.',
-		},
-		{
-			id: 'svg-transitions',
-			title: 'Transitions',
-			description:
-				'Build beautiful UIs with a powerful, performant transition engine built right into the framework.',
-		},
-	];
+	const examples = getContext('repl_widget_examples');
 
-	let selected = examples[0];
+	$: selected = examples[0];
 </script>
 
 <Section --background="radial-gradient(circle at 40% 30%, rgb(110, 113, 118), rgb(81, 93, 106))">
@@ -48,10 +25,12 @@
 			<a href="/examples">more <span class="large-show">&nbsp;examples</span> &rarr;</a>
 		</div>
 
-		<Example id={selected.id} />
+		{#if selected}
+			<Example id={selected?.id} />
+		{/if}
 	</div>
 
-	<p class="description">{@html selected.description}</p>
+	<p class="description">{@html selected?.description}</p>
 </Section>
 
 <style>
